# openapi.yaml
openapi: 3.0.3
info:
  title: Manufacturing Execution System API
  description: |
    Comprehensive MES API for jet engine component manufacturing.
    
    This API provides endpoints for managing work orders, quality control, 
    material tracking, equipment management, and comprehensive traceability 
    for aerospace manufacturing operations.
    
    ## Authentication
    All endpoints require JWT authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```
    
    ## Rate Limiting
    API requests are limited to 1000 requests per hour per user.
    
    ## Error Handling
    The API uses standard HTTP status codes. Error responses include:
    - `400` - Bad Request (validation errors)
    - `401` - Unauthorized (authentication required)
    - `403` - Forbidden (insufficient permissions)
    - `404` - Not Found
    - `409` - Conflict (business rule violation)
    - `422` - Unprocessable Entity (business validation error)
    - `500` - Internal Server Error
  version: 1.0.0
  contact:
    name: MES API Support
    email: support@mes-system.com
  license:
    name: Proprietary
servers:
  - url: https://api.mes-system.com/v1
    description: Production server
  - url: https://staging-api.mes-system.com/v1
    description: Staging server
  - url: http://localhost:3000/api/v1
    description: Development server
security:
  - BearerAuth: []
...
paths:
...
  /quality/inspections:
    get:
      tags:
        - Quality
      summary: List inspections
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: workOrderId
          in: query
          schema:
            type: string
            format: uuid
        - name: result
          in: query
          schema:
            type: string
            enum: [PASS, FAIL, CONDITIONAL]
        - name: dateFrom
          in: query
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Inspections retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Inspection'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Quality
      summary: Create inspection
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInspectionRequest'
      responses:
        '201':
          description: Inspection created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Inspection'
...
  /quality/ncrs:
    get:
      tags:
        - Quality
      summary: List non-conformance reports
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, UNDER_REVIEW, DISPOSITION_SET, CLOSED]
        - name: severity
          in: query
          schema:
            type: string
            enum: [MINOR, MAJOR, CRITICAL]
      responses:
        '200':
          description: NCRs retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/NonConformanceReport'
                  pagination:
                    $ref: '#/components/schemas/Pagination'

    post:
      tags:
        - Quality
      summary: Create non-conformance report
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateNCRRequest'
      responses:
        '201':
          description: NCR created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/NonConformanceReport'
...
  /equipment:
    get:
      tags:
        - Equipment
      summary: List equipment
      parameters:
        - name: siteId
          in: query
          schema:
            type: string
            format: uuid
        - name: workCenterId
          in: query
          schema:
            type: string
            format: uuid
        - name: status
          in: query
          schema:
            type: string
            enum: [AVAILABLE, BUSY, MAINTENANCE, DOWN]
      responses:
        '200':
          description: Equipment list retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Equipment'
  /equipment/{equipmentId}/status:
    get:
      tags:
        - Equipment
      summary: Get equipment status
      parameters:
        - name: equipmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Equipment status retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EquipmentStatus'

  # =====================================================
  # COMPONENT SCHEMAS
  # =====================================================
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    Pagination:
      type: object
      properties:
        page:
          type: integer
          minimum: 1
        limit:
          type: integer
          minimum: 1
        total:
          type: integer
          minimum: 0
        totalPages:
          type: integer
          minimum: 0
...
    Inspection:
      type: object
      properties:
        id:
          type: string
          format: uuid
        inspectionNumber:
          type: string
        workOrderOperationId:
          type: string
          format: uuid
        qualityPlanId:
          type: string
          format: uuid
        lotNumber:
          type: string
        serialNumber:
          type: string
        sampleSize:
          type: integer
        inspectorId:
          type: string
          format: uuid
        inspectionDate:
          type: string
          format: date-time
        overallResult:
          type: string
          enum: [PASS, FAIL, CONDITIONAL]
        notes:
          type: string
        certificateRequired:
          type: boolean
        certificateGenerated:
          type: boolean
        results:
          type: array
          items:
            $ref: '#/components/schemas/InspectionResult'
    InspectionResult:
      type: object
      properties:
        id:
          type: string
          format: uuid
        qualityCharacteristicId:
          type: string
          format: uuid
        characteristicName:
          type: string
        sampleNumber:
          type: integer
        measuredValue:
          type: number
        attributeValue:
          type: string
        result:
          type: string
          enum: [PASS, FAIL, NA]
        outOfSpecReason:
          type: string
        measurementEquipment:
          type: string
        measuredAt:
          type: string
          format: date-time
    CreateInspectionRequest:
      type: object
      required:
        - workOrderOperationId
        - qualityPlanId
        - sampleSize
        - inspectorId
      properties:
        workOrderOperationId:
          type: string
          format: uuid
        qualityPlanId:
          type: string
          format: uuid
        lotNumber:
          type: string
        serialNumber:
          type: string
        sampleSize:
          type: integer
          minimum: 1
        inspectorId:
          type: string
          format: uuid
...
    NonConformanceReport:
      type: object
      properties:
        id:
          type: string
          format: uuid
        ncrNumber:
          type: string
        inspectionId:
          type: string
          format: uuid
        workOrderId:
          type: string
          format: uuid
        partId:
          type: string
          format: uuid
        description:
          type: string
        quantityAffected:
          type: integer
        severity:
          type: string
          enum: [MINOR, MAJOR, CRITICAL]
        status:
          type: string
          enum: [OPEN, UNDER_REVIEW, DISPOSITION_SET, CLOSED]
        disposition:
          type: string
          enum: [REWORK, REPAIR, SCRAP, USE_AS_IS, RETURN_TO_SUPPLIER]
        dispositionReason:
          type: string
        rootCause:
          type: string
        correctiveAction:
          type: string
        reportedBy:
          type: string
          format: uuid
        reportedDate:
          type: string
          format: date-time
    CreateNCRRequest:
      type: object
      required:
        - partId
        - description
        - quantityAffected
        - severity
      properties:
        inspectionId:
          type: string
          format: uuid
        workOrderId:
          type: string
          format: uuid
        partId:
          type: string
          format: uuid
        description:
          type: string
        quantityAffected:
          type: integer
          minimum: 1
        severity:
          type: string
          enum: [MINOR, MAJOR, CRITICAL]
...
    Equipment:
      type: object
      properties:
        id:
          type: string
          format: uuid
        equipmentNumber:
          type: string
        name:
          type: string
        description:
          type: string
        workCenterId:
          type: string
          format: uuid
        manufacturer:
          type: string
        model:
          type: string
        serialNumber:
          type: string
        status:
          type: string
          enum: [AVAILABLE, BUSY, MAINTENANCE, DOWN]
        lastMaintenanceDate:
          type: string
          format: date
        nextMaintenanceDue:
          type: string
          format: date
    EquipmentStatus:
      type: object
      properties:
        equipmentId:
          type: string
          format: uuid
        status:
          type: string
          enum: [AVAILABLE, BUSY, MAINTENANCE, DOWN]
        currentWorkOrder:
          type: string
        currentOperation:
          type: string
        utilizationPercent:
          type: number
        lastStatusChange:
          type: string
          format: date-time
        maintenanceDue:
          type: boolean
        alerts:
          type: array
          items:
            type: object
            properties:
              type:
                type: string
              message:
                type: string
              severity:
                type: string
...