# src/routes/quality.ts
import express from 'express';
import { requireQualityAccess } from '../middleware/auth';
import { asyncHandler } from '../middleware/errorHandler';

const router = express.Router();

/**
 * @route GET /api/v1/quality/inspections
 * @desc Get inspections list
 * @access Private
 */
router.get('/inspections',
  requireQualityAccess,
  asyncHandler(async (req, res) => {
    // Mock response
    res.status(200).json({
      data: [],
      pagination: { page: 1, limit: 20, total: 0, totalPages: 0 }
    });
  })
);

/**
 * @route POST /api/v1/quality/inspections
 * @desc Create inspection
 * @access Private
 */
router.post('/inspections',
  requireQualityAccess,
  asyncHandler(async (req, res) => {
    // Mock response
    res.status(201).json({
      id: '1',
      message: 'Inspection created successfully'
    });
  })
);

/**
 * @route GET /api/v1/quality/ncrs
 * @desc Get non-conformance reports
 * @access Private
 */
router.get('/ncrs',
  requireQualityAccess,
  asyncHandler(async (req, res) => {
    // Mock response
    res.status(200).json({
      data: [],
      pagination: { page: 1, limit: 20, total: 0, totalPages: 0 }
    });
  })
);

export default router;